+++ src/libs/Mapbox.Maui/CameraChangedEventArgs.cs
@@ -1,0 +1,15 @@
+﻿namespace MapboxMaui;
+
+public class CameraChangedEventArgs : EventArgs
+{
+    public CameraChangedEventArgs()
+    {
+    }
+
+    public CameraChangedEventArgs(CameraOptions options)
+    {
+        Options = options;
+    }
+
+    public CameraOptions Options { get; }
+}
