+++ src/libs/Mapbox.Maui/IMapboxView.cs
@@ -30,6 +30,7 @@ public partial interface IMapboxView : IView
     Terrain Terrain { get; set; }
 
     Light Light { get; set; }
+    IViewAnnotationController ViewAnnotationController { get; }
 
     IAnnotationController AnnotationController { get; }
 
@@ -39,7 +40,6 @@ public partial interface IMapboxView : IView
 
     ILocationComponentPlugin LocationComponent { get; }
 }
-
 partial interface IMapboxView
 {
     event EventHandler MapReady;
@@ -68,85 +68,7 @@ partial interface IMapboxView
     ICommand RotatingCommand { get; }
 }
 
-public interface IAnnotationController
-{
-    IPolygonAnnotationManager CreatePolygonAnnotationManager(string id, LayerPosition layerPosition);
-    ICircleAnnotationManager CreateCircleAnnotationManager(string id, LayerPosition layerPosition);
-    IPointAnnotationManager CreatePointAnnotationManager(string id, LayerPosition layerPosition, ClusterOptions clusterOptions = null);
-    IPolylineAnnotationManager CreatePolylineAnnotationManager(string id, LayerPosition layerPosition);
-}
-
 public interface IMapFeatureQueryable
 {
     Task<IEnumerable<QueriedRenderedFeature>> QueryRenderedFeaturesWith(ScreenPosition point, RenderedQueryOptions options);
-}
-
-public interface IMapboxController
-{
-    IPosition GetMapPosition(ScreenPosition position);
-    CoordinateBounds GetCoordinateBoundsForCamera(CameraOptions cameraOptions);
-    ScreenPosition GetScreenPosition(IPosition position);
-    CameraOptions? CameraForCoordinates(
-        IEnumerable<IPosition> coordinates,
-        CameraOptions? cameraOptions = default,
-        Thickness? coordinatesPadding = default,
-        double? maxZoom = default,
-        ScreenPosition? offset = default
-    );
-    void CameraForCoordinates(
-        IEnumerable<IPosition> coordinates,
-        Action<CameraOptions?> completion,
-        CameraOptions? cameraOptions = default,
-        Thickness? coordinatesPadding = default,
-        double? maxZoom = default,
-        ScreenPosition? offset = default
-    );
-    void SetSourcePropertyFor<T>(string sourceId, string propertyName, T value, Action<Exception> completion = default);
-    void SetLayerPropertyFor<T>(string layerId, string propertyName, T value, Action<Exception> completion = default);
-}
-
-public class MapTappedEventArgs : EventArgs
-{
-    public MapTappedPosition Position { get; }
-
-    public MapTappedEventArgs(MapTappedPosition position)
-    {
-        Position = position;
-    }
-}
-public class CameraChangedEventArgs : EventArgs
-{
-    public CameraChangedEventArgs(CameraOptions options)
-    {
-        Options = options;
-    }
-
-    public CameraOptions Options { get; }
-}
-public class IndicatorAccuracyRadiusChangedEventArgs : EventArgs
-{
-    public IndicatorAccuracyRadiusChangedEventArgs(double radius)
-    {
-        Radius = radius;
-    }
-
-    public double Radius { get; }
-}
-public class IndicatorBearingChangedEventArgs : EventArgs
-{
-    public IndicatorBearingChangedEventArgs(double bearing)
-    {
-        Bearing = bearing;
-    }
-
-    public double Bearing { get; }
-}
-public class IndicatorPositionChangedEventArgs : EventArgs
-{
-    public IndicatorPositionChangedEventArgs(IPosition position)
-    {
-        Position = position;
-    }
-
-    public IPosition Position { get; }
 }
