+++ src/libs/Mapbox.Maui/IViewAnnotationController.cs
@@ -1,0 +1,6 @@
+﻿namespace MapboxMaui;
+
+public interface IViewAnnotationController
+{
+
+}
