+++ src/libs/Mapbox.Maui/IndicatorAccuracyRadiusChangedEventArgs.cs
@@ -1,0 +1,11 @@
+﻿namespace MapboxMaui;
+
+public class IndicatorAccuracyRadiusChangedEventArgs : EventArgs
+{
+    public IndicatorAccuracyRadiusChangedEventArgs(double radius)
+    {
+        Radius = radius;
+    }
+
+    public double Radius { get; }
+}
