+++ src/libs/Mapbox.Maui/IndicatorBearingChangedEventArgs.cs
@@ -1,0 +1,11 @@
+﻿namespace MapboxMaui;
+
+public class IndicatorBearingChangedEventArgs : EventArgs
+{
+    public IndicatorBearingChangedEventArgs(double bearing)
+    {
+        Bearing = bearing;
+    }
+
+    public double Bearing { get; }
+}
