+++ src/libs/Mapbox.Maui/IndicatorPositionChangedEventArgs.cs
@@ -1,0 +1,11 @@
+﻿namespace MapboxMaui;
+
+public class IndicatorPositionChangedEventArgs : EventArgs
+{
+    public IndicatorPositionChangedEventArgs(IPosition position)
+    {
+        Position = position;
+    }
+
+    public IPosition Position { get; }
+}
