+++ src/libs/Mapbox.Maui/MapTappedEventArgs.cs
@@ -1,0 +1,11 @@
+﻿namespace MapboxMaui;
+
+public class MapTappedEventArgs : EventArgs
+{
+    public MapTappedPosition Position { get; }
+
+    public MapTappedEventArgs(MapTappedPosition position)
+    {
+        Position = position;
+    }
+}
